+++ tests/expand/tests/expand/project_replace-enum.expanded.rs
@@ -1,11 +1,7 @@
 use pin_project::pin_project;
 #[pin(__private(project_replace))]
 enum Enum<T, U> {
-    Struct {
-        #[pin]
-        pinned: T,
-        unpinned: U,
-    },
+    Struct { #[pin] pinned: T, unpinned: U },
     Tuple(#[pin] T, U),
     Unit,
 }
@@ -33,10 +29,7 @@ enum __EnumProjectionRef<'pin, T, U>
 where
     Enum<T, U>: 'pin,
 {
-    Struct {
-        pinned: ::pin_project::__private::Pin<&'pin (T)>,
-        unpinned: &'pin (U),
-    },
+    Struct { pinned: ::pin_project::__private::Pin<&'pin (T)>, unpinned: &'pin (U) },
     Tuple(::pin_project::__private::Pin<&'pin (T)>, &'pin (U)),
     Unit,
 }
@@ -45,10 +38,7 @@ where
 #[allow(single_use_lifetimes)]
 #[allow(unreachable_pub)]
 enum __EnumProjectionOwned<T, U> {
-    Struct {
-        pinned: ::pin_project::__private::PhantomData<T>,
-        unpinned: U,
-    },
+    Struct { pinned: ::pin_project::__private::PhantomData<T>, unpinned: U },
     Tuple(::pin_project::__private::PhantomData<T>, U),
     Unit,
 }
@@ -58,40 +48,51 @@ enum __EnumProjectionOwned<T, U> {
 #[allow(clippy::used_underscore_binding)]
 const _: () = {
     impl<T, U> Enum<T, U> {
+        #[allow(dead_code)]
         fn project<'pin>(
             self: ::pin_project::__private::Pin<&'pin mut Self>,
         ) -> __EnumProjection<'pin, T, U> {
             unsafe {
                 match self.get_unchecked_mut() {
-                    Enum::Struct { pinned, unpinned } => __EnumProjection::Struct {
-                        pinned: ::pin_project::__private::Pin::new_unchecked(pinned),
-                        unpinned,
-                    },
-                    Enum::Tuple(_0, _1) => __EnumProjection::Tuple(
-                        ::pin_project::__private::Pin::new_unchecked(_0),
-                        _1,
-                    ),
+                    Enum::Struct { pinned, unpinned } => {
+                        __EnumProjection::Struct {
+                            pinned: ::pin_project::__private::Pin::new_unchecked(pinned),
+                            unpinned,
+                        }
+                    }
+                    Enum::Tuple(_0, _1) => {
+                        __EnumProjection::Tuple(
+                            ::pin_project::__private::Pin::new_unchecked(_0),
+                            _1,
+                        )
+                    }
                     Enum::Unit => __EnumProjection::Unit,
                 }
             }
         }
+        #[allow(dead_code)]
         fn project_ref<'pin>(
             self: ::pin_project::__private::Pin<&'pin Self>,
         ) -> __EnumProjectionRef<'pin, T, U> {
             unsafe {
                 match self.get_ref() {
-                    Enum::Struct { pinned, unpinned } => __EnumProjectionRef::Struct {
-                        pinned: ::pin_project::__private::Pin::new_unchecked(pinned),
-                        unpinned,
-                    },
-                    Enum::Tuple(_0, _1) => __EnumProjectionRef::Tuple(
-                        ::pin_project::__private::Pin::new_unchecked(_0),
-                        _1,
-                    ),
+                    Enum::Struct { pinned, unpinned } => {
+                        __EnumProjectionRef::Struct {
+                            pinned: ::pin_project::__private::Pin::new_unchecked(pinned),
+                            unpinned,
+                        }
+                    }
+                    Enum::Tuple(_0, _1) => {
+                        __EnumProjectionRef::Tuple(
+                            ::pin_project::__private::Pin::new_unchecked(_0),
+                            _1,
+                        )
+                    }
                     Enum::Unit => __EnumProjectionRef::Unit,
                 }
             }
         }
+        #[allow(dead_code)]
         fn project_replace(
             self: ::pin_project::__private::Pin<&mut Self>,
             __replacement: Self,
@@ -106,10 +107,14 @@ const _: () = {
                         };
                         let __guard = ::pin_project::__private::UnsafeOverwriteGuard {
                             target: __self_ptr,
-                            value: ::pin_project::__private::ManuallyDrop::new(__replacement),
+                            value: ::pin_project::__private::ManuallyDrop::new(
+                                __replacement,
+                            ),
                         };
                         {
-                            let __guard = ::pin_project::__private::UnsafeDropInPlaceGuard(pinned);
+                            let __guard = ::pin_project::__private::UnsafeDropInPlaceGuard(
+                                pinned,
+                            );
                         }
                         __result
                     }
@@ -120,10 +125,14 @@ const _: () = {
                         );
                         let __guard = ::pin_project::__private::UnsafeOverwriteGuard {
                             target: __self_ptr,
-                            value: ::pin_project::__private::ManuallyDrop::new(__replacement),
+                            value: ::pin_project::__private::ManuallyDrop::new(
+                                __replacement,
+                            ),
                         };
                         {
-                            let __guard = ::pin_project::__private::UnsafeDropInPlaceGuard(_0);
+                            let __guard = ::pin_project::__private::UnsafeDropInPlaceGuard(
+                                _0,
+                            );
                         }
                         __result
                     }
@@ -131,7 +140,9 @@ const _: () = {
                         let __result = __EnumProjectionOwned::Unit;
                         let __guard = ::pin_project::__private::UnsafeOverwriteGuard {
                             target: __self_ptr,
-                            value: ::pin_project::__private::ManuallyDrop::new(__replacement),
+                            value: ::pin_project::__private::ManuallyDrop::new(
+                                __replacement,
+                            ),
                         };
                         {}
                         __result
@@ -151,10 +162,10 @@ const _: () = {
         __field0: T,
         __field1: T,
     }
-    impl<'pin, T, U> ::pin_project::__private::Unpin for Enum<T, U> where
-        __Enum<'pin, T, U>: ::pin_project::__private::Unpin
-    {
-    }
+    impl<'pin, T, U> ::pin_project::__private::Unpin for Enum<T, U>
+    where
+        __Enum<'pin, T, U>: ::pin_project::__private::Unpin,
+    {}
     unsafe impl<T, U> ::pin_project::UnsafeUnpin for Enum<T, U> {}
     trait EnumMustNotImplDrop {}
     #[allow(clippy::drop_bounds, drop_bounds)]